chore: move copy-batch to amplify-cli-core
