+++ packages/amplify-cli-core/src/extensions/index.ts
@@ -1,2 +1,3 @@
 export * from './confirm-prompt';
 export * from './constants';
+export * from './copy-batch';
